Drop support for project USBify
The original idea was to have all of my configuration available on a
USB drive that would bootstrap itself when connected to a Mac. While
this is pretty cool from a Hollywood, hacker-porn standpoint, it is less
desirable to me due to its dependencies. Docker may be the better path
forward.