Adds emoji font and function to test emoji support
In my quest to learn more about terminals, I added a function to output ten
emojis. Technically this tests the same thing as test_unicode.

Unfortunately I couldn't get `st` to output any colored emojis. This is a bit of
a buzzkill for my grand plans to create a terminal-based chat client that
supports emojis.