Add daily Google ENG tips to `prodaccess`
Wraps the existing `prodaccess` executable and displays a quote from Google
ENG's fortune db.

Fortune is a GNU tool intended to support random quote compilation, display,
etc. It's pretty interesting.

NOTE: the `prodcertstatus` executable that this function is using as a guard
looks like it might be useful moving forward.