Remove opam-install variable and assertion
The `prelude/assert` for the existence of the `opam-install` directory was
failing.

I believe this assertion would have been failing sooner, but a bug in my
initialization was preventing Emacs from evaluating `wpc-ocaml.el`. It seems
that I removed whatever was jamming the initialization and as such, I uncovered
some more bugs.

Let this serve as a reminder that just because it hasn't bitten you yet, doesn't
mean that your software doesn't have a bug.